Use exclusive threads to optimize work wait queue
When a single work item is dispatched we only need to wake up one
thread to service it.  The current taskq implementation uses
non-exclusive processes in the work wait queue, so every thread in the
queue gets woken up when __taskq_dispatch() calls wake_up().  This
overhead can become non-negligible as the number of threads in the
queue increases, particularly when a large number of short work items
are being dispatched.  For example, this scenario appears to have led
to poor unlink performance in ZFS as described in Issues
zfsonlinux/zfs#458 and zfsonlinux/zfs#258.  To optimize this case, use
add_wait_queue_exclusive() to add threads to the work queue so only
one is woken up at a time.